Fixed NPE in WaterNode, added check for out-of-view shot source.

+++ src/java/com/threerings/bang/game/client/WaterNode.java
@@ -145,7 +145,9 @@ public class WaterNode extends Node
      */
     public void cleanup ()
     {
-        _smtstate.deleteAll();
+        if (_smtstate != null) {
+            _smtstate.deleteAll();
+        }
     }
     
     /**
